HelpMenu framework + CRASH LOG
Adding bare-bones functionality for an "Instructions" help menu option
at the main menu. Also adding a crashlog file: I'd left the main window
running for a while, gotten the "clamping to dt" console message, but
then out of nowhere the browser shut down and left this log. I'm
committing it so we have traceability if it happens again. (We wouldn't
have encountered it before because we rarely enabled the main menu, and
never stayed on it for very long.)